Moved component event queue, add interface
