+++ .clang-tidy
@@ -1,2 +1,2 @@
 Checks: ' -*, bugprone-*, cert-*, clang-analyzer-core.*, clang-analyzer-cplusplus.*, clang-analyzer-security.*, cppcoreguidelines-pro-bounds-pointer-arithmetic, cppcoreguidelines-pro-type-cstyle-cast, hicpp-* '
-WarningsAsErrors: 'bugprone-*,cert-*,clang-analyzer-*,hicpp-*'
+WarningsAsErrors: 'bugprone-*,cert-*,clang-analyzer-*
