+++ score/launch_manager/src/daemon/src/process_group_manager/details/BUILD
@@ -54,17 +54,42 @@ cc_library(
     visibility = ["//score/launch_manager/src/daemon/src/process_group_manager:__pkg__"],
     deps = [
         ":icomponent",
+        "//score/launch_manager/src/daemon/src/common:identifier_hash",
+    ]
+)
+
+cc_library(
+    name = "component_event_queue",
+    hdrs = ["component_event_queue.hpp"],
+    include_prefix = "score/mw/launch_manager/process_group_manager/details",
+    strip_include_prefix = "/score/launch_manager/src/daemon/src/process_group_manager/details",
+    visibility = ["//score/launch_manager/src/daemon/src/process_group_manager:__pkg__"],
+    deps = [
+        ":icomponent_event_receiver",
         "//score/launch_manager/src/daemon/src/common:constants",
         "//score/launch_manager/src/daemon/src/common:identifier_hash",
         "//score/launch_manager/src/daemon/src/common/concurrency:mpsc_bounded_queue",
+    ]
+)
+
+cc_library(
+    name = "icomponent_event_receiver",
+    hdrs = ["icomponent_event_receiver.hpp"],
+    include_prefix = "score/mw/launch_manager/process_group_manager/details",
+    strip_include_prefix = "/score/launch_manager/src/daemon/src/process_group_manager/details",
+    visibility = ["//score/launch_manager/src/daemon/src/process_group_manager:__pkg__"],
+    deps = [
+        ":component_event",
+        "//score/launch_manager/src/daemon/src/common:constants",
+        "//score/launch_manager/src/daemon/src/common:identifier_hash",
     ],
 )
 
 cc_test(
-    name = "component_event_UT",
-    srcs = ["component_event_UT.cpp"],
+    name = "component_event_queue_UT",
+    srcs = ["component_event_queue_UT.cpp"],
     deps = [
-        ":component_event",
+        ":component_event_queue",
         "@googletest//:gtest_main",
     ],
 )
@@ -115,7 +140,7 @@ cc_library(
     strip_include_prefix = "/score/launch_manager/src/daemon/src/process_group_manager/details",
     visibility = ["//score/launch_manager/src/daemon/src/process_group_manager:__pkg__"],
     deps = [
-        ":component_event",
+        ":component_event_queue",
         ":process_info_node",
         ":run_target",
         ":safe_process_map",
@@ -173,7 +198,7 @@ cc_library(
     strip_include_prefix = "/score/launch_manager/src/daemon/src/process_group_manager/details",
     visibility = ["//score/launch_manager/src/daemon/src/process_group_manager:__pkg__"],
     deps = [
-        ":component_event",
+        ":icomponent_event_receiver",
         ":os_handler",
         ":process_info_node",
         ":safe_process_map",
@@ -225,7 +250,7 @@ cc_library(
     }),
     visibility = ["//score/launch_manager/src/daemon/src/process_group_manager:__pkg__"],
     deps = [
-        ":component_event",
+        ":component_event_queue",
         ":graph",
         ":os_handler",
         ":process_monitor",
