+++ score/launch_manager/src/daemon/src/process_group_manager/details/component_event_queue_UT.cpp
@@ -11,7 +11,7 @@
  * SPDX-License-Identifier: Apache-2.0
  ********************************************************************************/
 
-#include "score/mw/launch_manager/process_group_manager/details/component_event.hpp"
+#include "score/mw/launch_manager/process_group_manager/details/component_event_queue.hpp"
 
 #include <gtest/gtest.h>
 #include <chrono>
