+++ score/launch_manager/src/daemon/src/process_group_manager/details/icomponent.hpp
@@ -16,7 +16,6 @@
 
 #include <score/stop_token.hpp>
 #include <cstdint>
-#include <functional>
 
 namespace score::lcm::internal
 {
