+++ score/launch_manager/src/daemon/src/process_group_manager/details/process_monitor.cpp
@@ -16,7 +16,7 @@
 namespace score::lcm::internal
 {
 
-ProcessMonitor::ProcessMonitor(ComponentEventQueue& event_queue) : event_queue_(event_queue) {}
+ProcessMonitor::ProcessMonitor(IComponentEventReceiver& event_queue) : event_queue_(event_queue) {}
 
 ProcessMonitor::~ProcessMonitor() = default;
